+++ robot_descriptions/j2s6s300_description.py
@@ -2,9 +2,8 @@
 # -*- coding: utf-8 -*-
 #
 # SPDX-License-Identifier: Apache-2.0
-# Copyright 2022 Stéphane Caron
 
-"""Kinova Gen3 description."""
+"""Kinova Jaco2 j2s6s300 description."""
 
 from os import getenv as _getenv
 from os import path as _path
@@ -12,14 +11,10 @@ from os import path as _path
 from ._cache import clone_to_cache as _clone_to_cache
 
 REPOSITORY_PATH: str = _clone_to_cache(
-    "ros2_kortex",
+    "kinova-ros",
     commit=_getenv("ROBOT_DESCRIPTION_COMMIT", None),
 )
 
-PACKAGE_PATH: str = _path.join(REPOSITORY_PATH, "kortex_description")
+PACKAGE_PATH: str = _path.join(REPOSITORY_PATH, "kinova_description")
 
-XACRO_PATH: str = _path.join(PACKAGE_PATH, "robots", "gen3.xacro")
-
-XACRO_ARGS = {
-    "dof": "7",
-}
+XACRO_PATH: str = _path.join(PACKAGE_PATH, "urdf", "j2s6s300_standalone.xacro")